requirements.txt
altgraph==0.17.4
packaging==24.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
speedtest-cli==2.1.3
